Added an enemy type, along with a rudimentary text display system
Added an enemy type, the circle. This was because at that time, the game was going to be different and procedurally generated, but that idea was scrapped. Its still not completely useless though.
Added a rudimentary text display system which changes based on which phase of the game the player is in.
Fixed some minor bugs with the gun of the player.